refactor(earn): replace upward graph icon
